+++ packages/xo-web/src/xo-app/settings/remotes/index.js
@@ -338,7 +338,10 @@ const INDIVIDUAL_ACTIONS = [
     level: 'primary',
   },
   {
-    handler: (remote, { editRemote }) => editRemote(remote),
+    handler: (remote, { reset, editRemote }) => {
+      reset()
+      editRemote(remote)
+    },
     icon: 'edit',
     label: _('formEdit'),
     level: 'primary',
@@ -398,6 +401,7 @@ export default decorate([
             columns={COLUMNS_LOCAL_REMOTE}
             data-editRemote={effects.editRemote}
             data-formatMessage={formatMessage}
+            data-reset={effects.reset}
             filters={FILTERS}
             groupedActions={GROUPED_ACTIONS}
             individualActions={INDIVIDUAL_ACTIONS}
@@ -414,6 +418,7 @@ export default decorate([
             columns={COLUMNS_NFS_REMOTE}
             data-editRemote={effects.editRemote}
             data-formatMessage={formatMessage}
+            data-reset={effects.reset}
             filters={FILTERS}
             groupedActions={GROUPED_ACTIONS}
             individualActions={INDIVIDUAL_ACTIONS}
@@ -430,6 +435,7 @@ export default decorate([
             columns={COLUMNS_SMB_REMOTE}
             data-editRemote={effects.editRemote}
             data-formatMessage={formatMessage}
+            data-reset={effects.reset}
             filters={FILTERS}
             groupedActions={GROUPED_ACTIONS}
             individualActions={INDIVIDUAL_ACTIONS}